+++ packages/gamut/src/Coachmark/index.tsx
@@ -1,0 +1,2 @@
+export * from './Coachmark';
+export * from './CoachmarkPopover';
